requirements.txt
PyMySQL==1.0.3
PyYAML~=6.0
pandas~=1.5.0